Remove graphVertexInfo.private.h header and split its content between the DFSUtils and Planarity private headers
