+++ c/graphLib/graphDFSUtils.private.h
@@ -14,7 +14,41 @@ extern "C"
 {
 #endif
 
-    // Currently nothing to add here.
+    /********************************************************************
+     Vertex Info Structure Definition.
+
+     This structure equips the non-virtual vertices with additional
+     information needed for DFS-related and planarity-related algorithms.
+
+        parent: The DFI of the DFS tree parent of this vertex
+        leastAncestor: min(DFI of neighbors connected by backedge)
+        lowpoint: min(leastAncestor, min(lowpoint of DFS Children))
+    */
+
+    struct vertexInfoRec
+    {
+        int parent, leastAncestor, lowpoint;
+
+        int visitedInfo;
+
+        int pertinentEdge,
+            pertinentRoots,
+            futurePertinentChild,
+            sortedDFSChildList,
+            fwdEdgeList;
+    };
+
+    typedef struct vertexInfoRec vertexInfoRec;
+    typedef vertexInfoRec *vertexInfoP;
+
+#define gp_GetVertexParent(theGraph, v) (theGraph->VI[v].parent)
+#define gp_SetVertexParent(theGraph, v, theParent) (theGraph->VI[v].parent = theParent)
+
+#define gp_GetVertexLeastAncestor(theGraph, v) (theGraph->VI[v].leastAncestor)
+#define gp_SetVertexLeastAncestor(theGraph, v, theLeastAncestor) (theGraph->VI[v].leastAncestor = theLeastAncestor)
+
+#define gp_GetVertexLowpoint(theGraph, v) (theGraph->VI[v].lowpoint)
+#define gp_SetVertexLowpoint(theGraph, v, theLowpoint) (theGraph->VI[v].lowpoint = theLowpoint)
 
 #ifdef __cplusplus
 }
